Modernize package and move to Python 3.14
