+++ analysis/README.md
@@ -1,0 +1,5 @@
+# Analysis
+
+Ancillary analysis code lives here. It is not installed with the Python
+package and may require tools or historical input files beyond the standard
+development environment.
