+++ artifacts/README.md
@@ -1,0 +1,8 @@
+# Historical artifacts
+
+These files are retained outputs from earlier simulation runs. They are not
+used by the package or test suite and do not define current published results.
+
+New ad hoc simulation output should remain untracked. Reproducible published
+results belong in `docs/` and must be generated through the scripts in
+`scripts/`.
