+++ experiments/README.md
@@ -1,0 +1,5 @@
+# Experiments
+
+This directory contains exploratory or legacy voting-method code that is not
+part of the supported `vse_sim` package. Production modules and tests must not
+depend on it.
