+++ mise.toml
@@ -1,2 +1,2 @@
 [tools]
-python = "3.12"
+python = "3.14"
