+++ pyproject.toml
@@ -1,13 +1,20 @@
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
 [project]
 name = "vse-sim"
 version = "0.1.0"
 description = "Voter Satisfaction Efficiency simulation tools"
-requires-python = ">=3.10,<3.13"
-dependencies = ["numpy>=1.23,<2", "scipy>=1.9,<1.12"]
+requires-python = ">=3.14,<3.15"
+dependencies = ["numpy>=2.3,<3", "scipy>=1.16,<2"]
 
 [dependency-groups]
-dev = ["matplotlib>=3.10,<4", "pytest>=7,<9"]
+dev = ["matplotlib>=3.10,<4", "pytest>=7,<10"]
+
+[tool.setuptools.packages.find]
+where = ["src"]
 
 [tool.pytest.ini_options]
 addopts = "--doctest-modules"
-testpaths = ["."]
+testpaths = ["src/vse_sim", "tests"]
