+++ src/vse_sim/__init__.py
@@ -1,0 +1,1 @@
+"""Voter Satisfaction Efficiency simulation tools."""
