--- test/test_binop.py
@@ -1,3 +1,0 @@
-def isnum(x):
-    """Test whether an object is an instance of a built-in numeric type."""
-    return next((1 for T in (int, float, complex) if isinstance(x, T)), 0)
